Solution for homework 4 in CMPS 367 Python Programming
Implements a well known algorithm in computer science commonly used in genetics research in both Python and C++ (Levenshtein Distance). Also evaluates the two, for both runtime efficiency and ease of programming. Purpose of thi is for to better understand the similarities, differences, and relative strengths of Python and C++ on a common problem.